orgnize code
